added example and moved index
